requirements.txt
fastapi
uvicorn
python-decouple
pydantic[email]
snowflake-snowpark-python